Add SoA red-black tree and compile-time reflection
This adds two major features on top of the structure-of-arrays (SoA) storage, plus the supporting refactors they needed.

Red-black tree (include/ball/types/rbtree.hpp, tests/case05_rbtree.cpp)
- New multi-column red-black tree stored in the SoA multivector: the ordering key and an arbitrary list of value columns each live in their own column, addressed by compile-time index (Get<N>) or by type.
- "No node"/end/no-child is an out-of-band index (-1), so there is no reserved sentinel row; real nodes start at slot 0.
- Erase compacts the array: the last node is moved into the vacated slot (index rewiring only, one payload move) and the row count shrinks, so storage stays dense with no holes. This keeps lookups, insert, erase and iteration at O(log n) and lets RootIndex() climb from the always-occupied first slot instead of scanning.
- Full API: Insert/Find/LowerBound/UpperBound/Contains, erase by key or iterator, bidirectional iterators, in-order foreach macros, and a Validate() that checks the whole RB topology and SoA links.
- Heap-backed and inline-buffer variants, plus set/map and single/multi value convenience aliases (RBTree_t, MultiRBTree_t, Buffer* spellings).
- The comparator is an empty base reached only through a Compare() accessor, so call sites never name the base or cast to it.
- Every method is annotated with its algorithmic complexity.

Compile-time reflection (include/ball/types/reflect.hpp, meta/reflect*.hpp, tests/case06_reflection.cpp)
- Field reflection via BALL_REFLECT_BEGIN/FIELD/END and BALL_REFLECT_BASE, exposing per-field type, name, offset and size.
- Standard-layout owners take a constant-expression offsetof fast path; non-standard-layout owners (e.g. those using reflected bases) measure offsets at run time from a pointer-to-member, avoiding ill-formed offsetof.
- A field-foreach that also exposes the padding before and after each field, threaded as values so it works for non-standard-layout owners too.
	- CReflect tagged wrappers give each value a distinct type even when the underlying type is shared, which is what lets the multivector address its columns by type.

SoA storage (multivector.hpp, viewbase.hpp, elements.hpp, elementspack.hpp)
- Per-column small-buffer-or-pointer union and bit-packed storage for small types (e.g. the 1-bit node color column).
- Block layout helpers (BlockOffsetBy/BlockSizeBy) and RelocateColumn, which picks memmove for trivially-relocatable types and move-construct + destruct otherwise.
- Fix SetCount on shrink: destruct the dropped rows in the current buffer before EnsureCapacity may reallocate to a smaller capacity. Doing it afterwards indexed past the shrunken column and, for the packed column, cleared a bit one byte into the next column. Erase never shrank storage before, so the tree's compaction is the first caller to hit this.
- AddToHead / AddMultipleToHead helpers.

Type traits and meta utilities (meta/)
- Replace the monolithic meta/traits.hpp with one header per trait (isclass, isunion, istrivial, istriviallycopyable, ismemmovesafe, is{default,copy,move}-constructible/assignable, istriviallydestructible, isstandardlayout, ...) and aggregate them in meta/typeinfo.hpp (MTypeInfo).
- ismemmovesafe drives the storage relocation choice.
- Add small meta helpers: tuple (typelist), indextype, indexsequence, size, and rework removecv/removereference and get.

Platform and macros
- Add c/nouniqueaddress.h (BALL_NO_UNIQUE_ADDRESS, routed to msvc::no_unique_address on MSVC/clang-cl) and c/constevaluated.h.
- Remove the now-unused c/memory.h.

Build/tests
- Wire Case05 (RB-tree) and Case06 (reflection) into the test runner and cmake test list; the RB-tree case includes a benchmark against std::map and std::unordered_map.

+++ src/ball/types/tests/case03_multivector.cpp
@@ -54,29 +54,29 @@ void Case03_MultiVector( TestsOutput_t &sOut )
 	funcAppendRange( 0, 10'000'000 );
 	funcCheck( "Initial fill" );
 
-	vec.AddToTail( size_t( 1 ), uint32_t( 11 ) );
+	vec.AddToTail( 1, 11 );
 	ref.push_back( { 1, 11 } );
-	vec.AddToTail( size_t( 2 ), uint32_t( 22 ) );
+	vec.AddToTail( 2, 22 );
 	ref.push_back( { 2, 22 } );
 	funcCheck( "AddToTail" );
 
-	vec.AddToHead( size_t( 0 ), uint32_t( 0 ) );
+	vec.AddToHead( 0, 0 );
 	ref.insert( ref.begin(), { 0, 0 } );
 	funcCheck( "AddToHead" );
 
-	vec.Insert( 1, size_t( 9 ), uint32_t( 99 ) );
+	vec.Insert( 1, 9, 99 );
 	ref.insert( ref.begin() + 1, { 9, 99 } );
 	funcCheck( "Insert single row" );
 
-	vec.InsertMultiple( 2, 2, size_t( 5 ), uint32_t( 50 ) );
+	vec.InsertMultiple( 2, 2, 5, 50 );
 	ref.insert( ref.begin() + 2, 2, TestPair_t{ 5, 50 } );
 	funcCheck( "InsertMultiple" );
 
-	vec.AddMultipleToHead( 2, size_t( 8 ), uint32_t( 80 ) );
+	vec.AddMultipleToHead( 2, 8, 80 );
 	ref.insert( ref.begin(), 2, TestPair_t{ 8, 80 } );
 	funcCheck( "AddMultipleToHead" );
 
-	vec.AddMultipleToTail( 3, size_t( 7 ), uint32_t( 70 ) );
+	vec.AddMultipleToTail( 3, 7, 70 );
 	ref.insert( ref.end(), 3, TestPair_t{ 7, 70 } );
 	funcCheck( "AddMultipleToTail" );
 
@@ -108,8 +108,7 @@ void Case03_MultiVector( TestsOutput_t &sOut )
 	{
 		for ( size_t i = 0; i < vec.Count(); ++i )
 		{
-			if ( vec.template At< size_t >( i ) == nFirst
-				&& vec.template At< uint32_t >( i ) == nSecond )
+			if ( vec.template At< size_t >( i ) == nFirst && vec.template At< uint32_t >( i ) == nSecond )
 			{
 				return i;
 			}
